legacy_flask/backend/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.35.1
Werkzeug==3.0.1
pymongo>=4.7.0
